+++ include/firebolt/discovery.h
@@ -47,7 +47,7 @@ public:
      *
      * @retval The status.
      */
-    virtual Result<bool> watched(const std::string& entityId, std::optional<float> progress,
+    virtual Result<bool> watched(const std::string& entityId, std::optional<double> progress,
                                  std::optional<bool> completed, std::optional<std::string> watchedOn,
                                  std::optional<AgePolicy> agePolicy) const = 0;
 };
